Add JavaScript functionality to toggle between two divs
This commit adds JavaScript code to toggle between two divs with the click of a button. The 'Click Me open' button in the 'a' div will hide the 'a' div and display the 'b' div, while the 'Close Me' button in the 'b' div will hide the 'b' div and display the 'a' div.